Suggestions from code review
